refactor: consolidate Shimmer import path and remove unused file
